+++ tools/autofix_lint.txt
@@ -52,3 +52,73 @@ No syntax errors detected in public/usermood.php
 =======
 ******* master
 ********* master
+# offset=0 2025-09-20T07:46:37+00:00
+
+Parse error: syntax error, unexpected identifier "name" in admin/class_promo.php on line 242
+Errors parsing admin/class_promo.php
+
+Parse error: syntax error, unexpected identifier "allComments" in admin/comments.php on line 25
+Errors parsing admin/comments.php
+No syntax errors detected in admin/namechanger.php
+
+Parse error: syntax error, unexpected end of file in admin/reports.php on line 120
+Errors parsing admin/reports.php
+No syntax errors detected in admin/reputation_ad.php
+No syntax errors detected in admin/shit_list.php
+No syntax errors detected in admin/sitelog.php
+No syntax errors detected in admin/system_view.php
+No syntax errors detected in admin/warn.php
+No syntax errors detected in include/function_account_delete.php
+No syntax errors detected in public/blackjack.php
+No syntax errors detected in public/coins.php
+No syntax errors detected in public/comment.php
+No syntax errors detected in public/credits.php
+No syntax errors detected in public/forums.php
+No syntax errors detected in public/friends.php
+No syntax errors detected in public/gift.php
+No syntax errors detected in public/invite.php
+No syntax errors detected in public/messages.php
+No syntax errors detected in public/reputation.php
+No syntax errors detected in public/staffbox.php
+No syntax errors detected in public/takeedit.php
+No syntax errors detected in public/takeeditcp.php
+No syntax errors detected in public/takereseed.php
+No syntax errors detected in public/topten.php
+No syntax errors detected in public/user_unlocks.php
+No syntax errors detected in public/usercp.php
+No syntax errors detected in public/userhistory.php
+No syntax errors detected in public/usermood.php
+
+Parse error: syntax error, unexpected string content "", expecting "-" or identifier or variable or number in public/users.php on line 117
+Errors parsing public/users.php
+# offset=0 2025-09-20T07:54:26+00:00 (retry)
+No syntax errors detected in admin/class_promo.php
+No syntax errors detected in admin/comments.php
+No syntax errors detected in admin/namechanger.php
+No syntax errors detected in admin/reports.php
+No syntax errors detected in admin/reputation_ad.php
+No syntax errors detected in admin/shit_list.php
+No syntax errors detected in admin/sitelog.php
+No syntax errors detected in admin/system_view.php
+No syntax errors detected in admin/warn.php
+No syntax errors detected in include/function_account_delete.php
+No syntax errors detected in public/blackjack.php
+No syntax errors detected in public/coins.php
+No syntax errors detected in public/comment.php
+No syntax errors detected in public/credits.php
+No syntax errors detected in public/forums.php
+No syntax errors detected in public/friends.php
+No syntax errors detected in public/gift.php
+No syntax errors detected in public/invite.php
+No syntax errors detected in public/messages.php
+No syntax errors detected in public/reputation.php
+No syntax errors detected in public/staffbox.php
+No syntax errors detected in public/takeedit.php
+No syntax errors detected in public/takeeditcp.php
+No syntax errors detected in public/takereseed.php
+No syntax errors detected in public/topten.php
+No syntax errors detected in public/user_unlocks.php
+No syntax errors detected in public/usercp.php
+No syntax errors detected in public/userhistory.php
+No syntax errors detected in public/usermood.php
+No syntax errors detected in public/users.php
