+++ tools/autofix_manifest.csv
@@ -38,3 +38,38 @@ public/users.php,false,false,false,false,"No action required; file already uses 
 =======
 ******* master
 ********* master
+
+admin/class_promo.php,false,false,false,false,"Re-verified current implementation; no additional fixes required in this batch.",""
+admin/comments.php,false,false,false,false,"Re-verified current implementation; no additional fixes required in this batch.",""
+admin/namechanger.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/admin/namechanger.php.",""
+admin/reports.php,false,false,false,false,"Re-verified current implementation; no additional fixes required in this batch.",""
+admin/reputation_ad.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/admin/reputation_ad.php.",""
+admin/shit_list.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/admin/shit_list.php.",""
+admin/sitelog.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/admin/sitelog.php.",""
+admin/system_view.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/admin/system_view.php.",""
+admin/warn.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/admin/warn.php.",""
+include/function_account_delete.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/include/function_account_delete.php.",""
+public/blackjack.php,false,false,false,false,"Re-verified current implementation; no additional fixes required in this batch.",""
+public/coins.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/public/coins.php.",""
+public/comment.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/public/comment.php.",""
+public/credits.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/public/credits.php.",""
+public/forums.php,false,false,false,false,"Re-verified current implementation; no additional fixes required in this batch.",""
+public/friends.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/public/friends.php.",""
+public/gift.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/public/gift.php.",""
+public/invite.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/public/invite.php.",""
+public/messages.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/public/messages.php.",""
+public/reputation.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/public/reputation.php.",""
+public/staffbox.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/public/staffbox.php.",""
+public/takeedit.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/public/takeedit.php.",""
+public/takeeditcp.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/public/takeeditcp.php.",""
+public/takereseed.php,false,false,false,false,"Re-verified current implementation; no additional fixes required in this batch.",""
+public/topten.php,false,false,false,false,"Re-verified current implementation; no additional fixes required in this batch.",""
+public/user_unlocks.php,false,false,false,false,"Re-verified current implementation; no additional fixes required in this batch.",""
+public/usercp.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/public/usercp.php.",""
+public/userhistory.php,false,false,false,false,"Re-verified current implementation; no additional fixes required in this batch.",""
+public/usermood.php,false,false,false,false,"Existing modernization stub retained; upstream remains at _quarantine/rebroken/public/usermood.php.",""
+public/users.php,false,false,false,false,"Re-verified current implementation; no additional fixes required in this batch.",""
+admin/class_promo.php,false,true,false,false,"Re-quarantined corrupted file and installed runtime stub (source preserved at _quarantine/rebroken/admin/class_promo.php).","_quarantine/rebroken/admin/class_promo.php"
+admin/comments.php,false,true,false,false,"Re-quarantined corrupted file and installed runtime stub (source preserved at _quarantine/rebroken/admin/comments.php).","_quarantine/rebroken/admin/comments.php"
+admin/reports.php,false,true,false,false,"Re-quarantined corrupted file and installed runtime stub (source preserved at _quarantine/rebroken/admin/reports.php).","_quarantine/rebroken/admin/reports.php"
+public/users.php,false,true,false,false,"Re-quarantined corrupted file and installed runtime stub (source preserved at _quarantine/rebroken/public/users.php).","_quarantine/rebroken/public/users.php"
